requirements.txt
nodriver
motor
python-dotenv
pydantic>=2.0
